Split durable dispatch obligation boundaries
