+++ src/mindroom/dispatch_obligations/__init__.py
@@ -1,0 +1,13 @@
+"""Durable exact obligations for fallible Matrix callback dispatch."""
+
+from .events import callback_kind_for_source_kind
+from .runner import DispatchObligationRunner
+from .storage import DispatchCallbackKind, DispatchObligationStore, DispatchSemanticConsumer
+
+__all__ = [
+    "DispatchCallbackKind",
+    "DispatchObligationRunner",
+    "DispatchObligationStore",
+    "DispatchSemanticConsumer",
+    "callback_kind_for_source_kind",
+]
