+++ tach.toml
@@ -92,7 +92,7 @@ path = "mindroom.dispatch_callback_outcome"
 depends_on = []
 visibility = [
     "mindroom.bot",
-    "mindroom.dispatch_obligations",
+    "mindroom.dispatch_obligations.events",
     "mindroom.turn_controller",
 ]
 
@@ -101,7 +101,7 @@ path = "mindroom.dispatch_recovery_context"
 depends_on = []
 visibility = [
     "mindroom.coalescing",
-    "mindroom.dispatch_obligations",
+    "mindroom.dispatch_obligations.runner",
     "mindroom.turn_controller",
     "mindroom.visible_voice_echo",
 ]
@@ -109,16 +109,47 @@ visibility = [
 [[modules]]
 path = "mindroom.dispatch_obligations"
 depends_on = [
-    "mindroom.background_tasks",
-    "mindroom.dispatch_callback_outcome",
-    "mindroom.dispatch_recovery_context",
-    "mindroom.logging_config",
-    "mindroom.matrix.media",
+    "mindroom.dispatch_obligations.events",
+    "mindroom.dispatch_obligations.runner",
+    "mindroom.dispatch_obligations.storage",
 ]
 visibility = [
     "mindroom.bot",
     "mindroom.reaction_dispatch",
 ]
+
+[[modules]]
+path = "mindroom.dispatch_obligations.storage"
+depends_on = ["mindroom.logging_config"]
+visibility = [
+    "mindroom.dispatch_obligations",
+    "mindroom.dispatch_obligations.events",
+    "mindroom.dispatch_obligations.runner",
+]
+
+[[modules]]
+path = "mindroom.dispatch_obligations.events"
+depends_on = [
+    "mindroom.dispatch_callback_outcome",
+    "mindroom.dispatch_obligations.storage",
+    "mindroom.dispatch_source",
+    "mindroom.matrix.media",
+]
+visibility = [
+    "mindroom.dispatch_obligations",
+    "mindroom.dispatch_obligations.runner",
+]
+
+[[modules]]
+path = "mindroom.dispatch_obligations.runner"
+depends_on = [
+    "mindroom.background_tasks",
+    "mindroom.dispatch_obligations.events",
+    "mindroom.dispatch_obligations.storage",
+    "mindroom.dispatch_recovery_context",
+    "mindroom.logging_config",
+]
+visibility = ["mindroom.dispatch_obligations"]
 
 [[modules]]
 path = "mindroom.reaction_dispatch"
